@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "edu.cs4730.facetrackermultidemo"
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 19
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -20,8 +20,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.google.android.gms:play-services-vision:15.0.0'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    //not needed, but removes a issue with appcompat versions.
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'com.google.android.gms:play-services-vision:17.0.2'
+    implementation 'com.google.android.gms:play-services-base:16.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 }
